Initial Commit
Looking into the Indev file format! The game didn't have a built-in upgrade path going from Indev to Infdev, so I'm going to try and build one as well. BJMastermind Games made a program in Java to do it, so thanks to him I'm looking towards his work as inspiration on how to possibly accomplish it.

The first part I'm starting with is making TS types for the world format, since that will only make things easier to work with when writing the conversion logic itself. I'll still have to do that for Infdev as well.

https://github.com/BJTMastermind/Indev-Updater
https://minecraft.wiki/w/Java_Edition_data_values/Indev
https://minecraft.wiki/w/Java_Edition_Indev_level_format
https://minecraft.wiki/w/Java_Edition_data_values#Entities
https://minecraft.wiki/w/Java_Edition_history_of_textures/Entities
https://minecraft.wiki/w/Java_Edition_data_values?oldid=417712#Entity_IDs
https://minecraft.wiki/w/Java_Edition_history_of_textures/Paintings
https://minecraft.wiki/w/Java_Edition_16w32a#General_2