World Update (Lava Wave)
Kind of made an oopsie, forgot about the old Lava spreading mechanics. Well, "forgot" hehe.

And turns out the Lit Furnace variant is only in an item form, the Tile Entity itself, is still the same as a regular Furnace. So you have to check the Block ID in the world (chunk eventually) data to see which variant it actually is. I thought this concept was something that didn't make sense, I discovered it over in Bedrock with the Beehive / Bee Nest conundrum, and I thought it was a Bedrock wonk. But, looks like it's been a kind of pattern that the game has been using for a while!