Types Organization + SNBT Errors?
Going to properly move things into their own types files/modules/folders, and the like.

Was going to mention, I had to write a script to do the NBT editing, rather than just use SNBT, because it looks like my SNBT parser currently doesn't handle (not sure which it is) either really long lists, which breaks the rest of the parsing, or the use of bigger numbers inside of a List tag, say like highly-precise Float tag values. Something about the SNBT parser reading those isn't working right, and I have to figure out why reading it as SNBT isn't working out correctly. It's not NBTify's binary NBT implementation, because that can read and write the same file no issue, it's just going from the SNBT output back to JS that seems to error. The JS to SNBT output also seems to be working great, so I don't know of any issues here.

https://github.com/Offroaders123/NBTify/issues/45