Simpler Types Modules
Making things more similar to like how I manage the types over in Region-Types.

Things are starting to make more sense in terms of how I may be able to write types for specific platform versions, and use types for each of those as needed. Say like how these could be for Indev and Infdev, and any additive changes can be combined/derived from one to the other. When a release has breaking changes in NBT shapes and resource IDs (like The Flattening), then that can be a new base format version, where other versions onwards as derived from. Say like how 1.14 could use 1.13's types, but additive. Things could be broken into feature sets as well, like how the Village features changed from 1.13 to 1.14. So the base types for building major platform versions, those can be built from reusable types that span multiple versions, per se. I need to have all of the versions typed out in order to find those similarities and differences though. That's more of an after-the-fact thing to deduce, rather than trying to plan out those modules ahead of time.

Been listening to my new music while working on this, and I'm starting to like it! Maybe it's the fun with working on this that helps haha.

Glad the time has come where working with React feels normal now! Tried moving Dovetail from SolidJS to React today, and things like state management came easy to me, in knowing how Solid works vs React. Reading about the components' lifecycles really helps it click, at least that seemed to be part of the hangup for me. It's that in React, the body of the function runs the whole thing every time. Working with SolidJS, I learned to understand where to keep track of reactive endpoints, which I think it's the big hurdle I was having. I will fully recommend learning Solid before React because of that, because you get more of a concept of how all of the different parts tie together. Let alone, it being more performant hehe. Solid for the win! I want to get React on my 'resume' though hahaha.